requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Django==3.2.9
firebase-admin==5.1.0
gcloud==0.17.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
jws==0.1.3
msgpack==1.0.2
mysqlclient==2.1.0
oauth2client==3.0.0
outcome==1.1.0
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
pyOpenSSL==21.0.0
pyparsing==3.0.6
Pyrebase==3.0.27
python-decouple==3.5
python-jwt==2.0.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.7.0
rsa==4.7.2
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.7
wsproto==1.0.0
